More polishing the float-double thing. New rule: floats if we store a lot of them - e.g. in vectors - and doubles if we're just doing math - e.g. time.

+++ include/WorldObjects/Lighting/Light.hpp
@@ -11,7 +11,7 @@ public:
     Light() = default;
     virtual ~Light() override { glDisable(m_lightid); }
 
-    virtual void update(double t, float dt) override;
+    virtual void update(double t, double dt) override;
     void enable();
 
     Color ambient() { return m_ambient; }
